Refactor cli to stand-alone module; Update cli entry point in setup.py
